Add StrongInterner. Move and refactor WorldTexture implementation.
