+++ chunky/src/java/se/llbit/chunky/renderer/scene/biome/BiomeStructure.java
@@ -1,8 +1,8 @@
 package se.llbit.chunky.renderer.scene.biome;
 
 import it.unimi.dsi.fastutil.objects.Object2ReferenceOpenHashMap;
+import se.llbit.chunky.renderer.scene.biome.worldtexture.WorldTexture2dBiomeStructure;
 import se.llbit.chunky.world.Chunk;
-import se.llbit.chunky.world.WorldTexture;
 import se.llbit.log.Log;
 import se.llbit.math.structures.Position2IntStructure;
 import se.llbit.math.structures.Position2ReferenceStructure;
@@ -28,9 +28,6 @@ public interface BiomeStructure extends Position2ReferenceStructure<float[]> {
   }
 
   /**
-   * This is basically a reimplementation of {@link WorldTexture#load} but instead loading into an arbitrary
-   * BiomeStructure implementation
-   *
    * @param impl The implementation to load the legacy implementation into
    * @param in   The serialised legacy data in an input stream
    * @return The newly constructed {@link BiomeStructure} of the specified implementation
